+++ webapp/.env
@@ -2,12 +2,6 @@ NEXT_PUBLIC_BTC_INPUTS_SIZE=105
 NEXT_PUBLIC_BTC_OUTPUTS_SIZE=25
 # Once mainnet is published, these values below may go into .env.development as they belong to hemi's testnet
 NEXT_PUBLIC_MEMPOOL_API_URL="https://mempool.space/testnet/api"
-NEXT_PUBLIC_ADDRESS_MANAGER="0x23f0022354241FDb721Dc43E7897d7Af662A2995"
-NEXT_PUBLIC_L2_BRIDGE="0x4200000000000000000000000000000000000010"
-NEXT_PUBLIC_L2_OUTPUT_ORACLE_PROXY="0x032d1e1dd960A4B027a9a35FF8B2b672E333Bc27"
-NEXT_PUBLIC_OPTIMISM_PORTAL_PROXY="0xB6f9579980aE46f61217A99145645341E49E2516"
-NEXT_PUBLIC_PROXY_OVM_L1_CROSS_DOMAIN_MESSENGER="0x9bCCCf1d222539c4C47E4C6f5749e4d5fA33215c"
-NEXT_PUBLIC_PROXY_OVM_L1_STANDARD_BRIDGE="0xc94b1BEe63A3e101FE5F71C80F912b4F4b055925"
 NEXT_PUBLIC_RAINBOW_APP_NAME=Hemi
 NEXT_PUBLIC_RAINBOW_PROJECT_ID=THE_PROJECT_ID
 NEXT_PUBLIC_TESTNET_MODE=true
