+++ webapp/hooks/useL2Bridge.ts
@@ -14,60 +14,14 @@ import { hemi } from 'app/networks'
 import { useJsonRpcProvider, useWeb3Provider } from 'hooks/useEthersSigner'
 import { useIsConnectedToExpectedNetwork } from 'hooks/useIsConnectedToExpectedNetwork'
 import { Token } from 'types/token'
-import { type Address, type Chain, type Hash, isHash, zeroAddress } from 'viem'
+import {
+  getCrossChainMessenger,
+  getTunnelContracts,
+} from 'utils/crossChainMessenger'
+import { type Chain, type Hash, isHash } from 'viem'
 import { useAccount } from 'wagmi'
 
 import { useEstimateFees } from './useEstimateFees'
-
-const sdkPromise = import('@eth-optimism/sdk')
-
-const l1Contracts = {
-  AddressManager: process.env.NEXT_PUBLIC_ADDRESS_MANAGER as Address,
-  BondManager: zeroAddress,
-  CanonicalTransactionChain: zeroAddress,
-  L1CrossDomainMessenger: process.env
-    .NEXT_PUBLIC_PROXY_OVM_L1_CROSS_DOMAIN_MESSENGER as Address,
-  L1StandardBridge: process.env
-    .NEXT_PUBLIC_PROXY_OVM_L1_STANDARD_BRIDGE as Address,
-  L2OutputOracle: process.env.NEXT_PUBLIC_L2_OUTPUT_ORACLE_PROXY as Address,
-  OptimismPortal: process.env.NEXT_PUBLIC_OPTIMISM_PORTAL_PROXY as Address,
-  StateCommitmentChain: zeroAddress,
-}
-
-async function getCrossChainMessenger({
-  l1ChainId,
-  l1Signer,
-  l2Signer,
-}: {
-  l1ChainId: Chain['id']
-  l1Signer: SignerOrProviderLike
-  l2Signer: SignerOrProviderLike
-}) {
-  const { CrossChainMessenger, ETHBridgeAdapter, StandardBridgeAdapter } =
-    await sdkPromise
-  return new CrossChainMessenger({
-    bedrock: true,
-    bridges: {
-      ETH: {
-        Adapter: ETHBridgeAdapter,
-        l1Bridge: l1Contracts.L1StandardBridge,
-        l2Bridge: process.env.NEXT_PUBLIC_L2_BRIDGE,
-      },
-      Standard: {
-        Adapter: StandardBridgeAdapter,
-        l1Bridge: l1Contracts.L1StandardBridge,
-        l2Bridge: process.env.NEXT_PUBLIC_L2_BRIDGE,
-      },
-    },
-    contracts: {
-      l1: l1Contracts,
-    },
-    l1ChainId,
-    l1SignerOrProvider: l1Signer,
-    l2ChainId: hemi.id,
-    l2SignerOrProvider: l2Signer,
-  })
-}
 
 type GasEstimationOperations = Extract<
   keyof CrossChainMessengerType['estimateGas'],
@@ -374,7 +328,7 @@ export const useDepositErc20Token = function ({
     depositErc20TokenError,
     depositErc20TokenGasFees,
     depositErc20TokenTxHash,
-    l1StandardBridgeAddress: l1Contracts.L1StandardBridge,
+    l1StandardBridgeAddress: getTunnelContracts(l1ChainId).L1StandardBridge,
     resetDepositToken,
     status,
   }
