+++ webapp/package.json
@@ -21,7 +21,7 @@
     "btc-wallet": "1.0.0",
     "camelcase-keys": "7.0.2",
     "hemi-socials": "1.0.0",
-    "hemi-viem": "1.3.0",
+    "hemi-viem": "1.5.0",
     "javascript-time-ago": "2.5.10",
     "next": "14.2.3",
     "next-intl": "2.22.1",
